refactoring other classes
